% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base_p}
\alias{base_p}
\title{Extract the base of a node name}
\usage{
base_p(p)
}
\arguments{
\item{p}{character vector containing node names}
}
\description{
Removes square brackets, numbers, and commas that represent
the index of the node element in question. Returns just the
text portion. E.g, \code{"R\[1\]"} becomes \code{"R"}.
}
\note{
this is NOT a function users will generally use.
}
