% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_post.R
\name{bind_post}
\alias{bind_post}
\title{Combine two mcmc.list objects}
\usage{
bind_post(post1, post2)
}
\arguments{
\item{post1}{an object of class \code{mcmc.list}}

\item{post2}{an object of class \code{mcmc.list}}
}
\description{
Take two mcmc.list objects, possibly from the same or different models.
The two posterior objects must have equal chains and iterations per chain.
}
\details{
The two lists are coerced to matricies, \code{cbind()}-ed, and
  converted back to to an \code{mcmc.list} using \code{matrix2mcmclist}.
  The chains, thinning, and iterations, burn-in settings in the output object
  will remain as for \code{post1}.
}
