% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cjs}
\alias{cjs}
\title{Example mcmc.list 1}
\format{
A \code{\link[coda:mcmc.list]{mcmc.list}} object.
}
\source{
Posterior samples generated from a model fitted to hypothetical data set.
See \code{vignette("example-mcmclists")} on the context, model, and monitored parameters.
}
\usage{
cjs
}
\description{
An example of samples from a joint posterior distribution from a Cormack-Jolly-Seber model.
\strong{The specific context does not matter, this object is provided to show examples of 'postpack' functionality}.
}
\keyword{datasets}
