% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_index}
\alias{drop_index}
\title{Extract the base node name of a parameter}
\usage{
drop_index(params)
}
\arguments{
\item{params}{A character vector with length >= 1 containing node names.}
}
\value{
A character vector with the same length as \code{params}, with no indices included.
}
\description{
Removes square brackets, numbers, and commas that represent
the index of the node element in question. Returns just the node name.
}
\note{
This is NOT a function users will generally use.
}
\examples{
postpack:::drop_index("a[1]")
postpack:::drop_index(c("a[1]", "a[2]"))
postpack:::drop_index(c("a[1,1,1,1,1]", "a[2,2,2,2,2]"))
}
