% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_post.R
\name{filter_post}
\alias{filter_post}
\title{Extract desired nodes from mcmc.list}
\usage{
filter_post(post, p = NULL, format = "mcmc.list", iters = F,
  chains = F)
}
\arguments{
\item{post}{an object of class \code{mcmc.list} from which to filter out the node
identified by \code{p}}

\item{p}{a character vector of with length >=. Passed to \code{stringr::str_detect()},
so can (and perhaps should) be a regular expression.}

\item{format}{a character vector of length 1 specifying the desired output format.
Currently accepted options are \code{format = "mcmc.list"} and \code{format = "matrix"}.
Defaults to \code{format = "mcmc.list"}.}

\item{iters}{logical. Do you wish to retain the iteration number if \code{format = "mcmc.list"}?}

\item{chains}{logical. Do you wish to retain the chain number if \code{format = "mcmc.list"}?}
}
\description{
In many cases, it is helpful to perform operations on smaller subsets of a
\code{mcmc.list} object generated by a package like \code{jagsUI} or \code{coda}.
This function filters out a smaller subset corresponding only to the node requested.
It can be seen as one of the the "starting points" of the workflow of the functions in this package.
}
