% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ins_regex_bracket}
\alias{ins_regex_bracket}
\title{Insert escapes on regex brackets}
\usage{
ins_regex_bracket(string)
}
\arguments{
\item{string}{a character vector}
}
\description{
Insert escapes on regex brackets
}
\details{
searches the contents of a string for the occurrence of a
 square bracket or two, and inserts the necessary escapes for searching
 via regular expression
}
\note{
this is NOT a function users will generally use.
}
