% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ins_regex_lock}
\alias{ins_regex_lock}
\title{Insert the symbols to lock in a string for matching}
\usage{
ins_regex_lock(params)
}
\arguments{
\item{params}{A character vector with length >= 1 to paste a \code{^} and \code{$}
(if not already present) to lock in the match.}
}
\value{
A character vector with locking anchors inserted to force an exact match. For example,
\verb{"a\\\[1\\\]"} becomes \verb{"^a\\\[1\\\]$"}.
}
\description{
To ensure that a regular expression will match exactly,
it's necessary to specify so.
}
\note{
This is \strong{not} a function users will generally use directly.
}
