% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_out}
\alias{list_out}
\title{List vector elements in a nice format}
\usage{
list_out(x, final = NULL, per_line = 1e+06, wrap = NULL, indent = NULL)
}
\arguments{
\item{x}{A vector, will be coerced to a character.}

\item{final}{Word that will separate the final element in the list from others.
See the examples.}

\item{per_line}{Number of elements printed per line. See the examples.}

\item{wrap}{Optional character to wrap around each element, e.g., quotation marks.}

\item{indent}{Optional string to place in front of the first
element on each line. See the examples.}
}
\value{
A character vector with length == 1; ready to be passed to
\code{\link[base:stop]{base::stop()}}, \code{\link[base:warning]{base::warning()}}, or \code{\link[base:cat]{base::cat()}}, to provide a useful message.
}
\description{
Converts a vector into a comma-separated list for use in sentences (error messages, warnings, etc.).
}
\examples{
postpack:::list_out(x = letters[1:10])
cat(postpack:::list_out(letters[1:26], wrap = "'", per_line = 4, final = "or", indent = "  "))
}
