% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_p.R
\name{match_p}
\alias{match_p}
\title{Find matching node names}
\usage{
match_p(post, p, ubase = F, warn = F)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{p}{a character vector with length >= 1. Passed to \code{stringr::str_detect},
so can, and sometimes should, be a regular expression; see the examples.
Duplicate matches found among different elements of \code{p} are discarded.}

\item{ubase}{logical. Do you wish to return only the unique bases (i.e., without indices listed)?}

\item{warn}{logical. Do you wish to be warned when duplicates and cases
with more than one match per \code{p} element? Defaults to \code{FALSE}, and is provided for
helping to diagnose problems with regex matching.}
}
\value{
a character vector with all node names that match \code{p}.
  If no matches are found, it will return an error with
  the base node names found the \code{mcmc.list} to help the next try.
}
\description{
Returns all the node names stored in an \code{mcmc.list} object
that match a provided string
}
\details{
this function is called as one of the first steps in many of the more downstream
  functions in this package. It is thus fairly important to get used to
  how the regular expressions work, particularly
  with the \code{\link[stringr]{str_detect}} function.
}
\examples{
\dontrun{
match_p(post, "alpha")
match_p(post, c("S", "R"))   # Elements of  S, S_msy, and Sigma_R, and R
match_p(post, c("R["))       # Elements of R and Sigma_R only
match_p(post, c("^R["))      # Elements of R
}

}
