% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mytitle}
\alias{mytitle}
\title{Adds a title between two figures}
\usage{
mytitle(text)
}
\arguments{
\item{text}{character vector length == 1. The text string to
include as a title}
}
\description{
Used by \code{\link{diag_plots}} to place a common
title over top of two figures: one trace and one density
for a given node.
}
\note{
this is NOT a function users will generally use.
}
