% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_subset.R
\name{post_subset}
\alias{post_subset}
\title{Extract a subset of nodes from a \code{mcmc.list}}
\usage{
post_subset(
  post,
  params,
  matrix = FALSE,
  iters = FALSE,
  chains = FALSE,
  auto_escape = TRUE
)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{params}{a character vector of with length >= 1 specifying the nodes to extract. Passed to \code{\link{match_params}},
so can (and sometimes should) be a regular expression.}

\item{matrix}{logical. If \code{TRUE}, the subsetted output will be returned as a matrix.
Defaults to \code{FALSE}, in which case the class \code{mcmc.list} will be retained.}

\item{iters}{logical. Do you wish to retain the iteration number if \code{matrix = TRUE}? Not used otherwise.
Defaults to \code{FALSE}.}

\item{chains}{logical. Do you wish to retain the chain number if \code{matrix = TRUE}? Not used otherwise.
Defaults to \code{FALSE}.}

\item{auto_escape}{logical. \code{FALSE} will treat \code{"["} and \code{"]"}
as regular expression syntax (unless explicitly escaped by user),
\code{TRUE} will treat these symbols as plain text to be matched.
It is generally recommended to keep this as \code{TRUE} (the default),
unless you are performing complex regex searches that require the
\code{"["} and \code{"]"} symbols to be special characters}
}
\description{
Subsets a smaller portion from the joint posterior
corresponding only to the node(s) requested.
}
\examples{
# load example mcmc.list
data(cjs, package = "postpack")

# create mcmc.list with all nodes that contain "B0"
x1 = post_subset(cjs, "B0")

# create mcmc.list with all nodes that contain "b" or "B"
x2 = post_subset(cjs, c("b", "B"))

# perform the subset and return a matrix as output, while retaining the chain ID
x3 = post_subset(cjs, "B0", matrix = T, chain = T)
}
\seealso{
\code{\link{match_params}}
}
