% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_summ.R
\name{post_summ}
\alias{post_summ}
\title{Obtain a posterior summary of specific nodes}
\usage{
post_summ(
  post,
  params,
  digits = NULL,
  probs = c(0.5, 0.025, 0.975),
  Rhat = FALSE,
  neff = FALSE,
  mcse = FALSE,
  by_chain = F,
  auto_escape = TRUE
)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{params}{a character vector of with length >= 1 specifying the nodes to summarize.
Passed to \code{\link{match_params}}, so can (and sometimes should) be a regular expression.}

\item{digits}{numeric vector controlling rounding of summaries.
Passed to \code{\link[base]{round}} and defaults to \code{NULL}, which produces no rounding.}

\item{probs}{numeric vector specifying the posterior quantiles you wish to have summarized.
Passed to \code{\link[stats]{quantile}}.
Defaults to \code{probs = c(0.5, 0.025, 0.975)} (i.e., median and equal-tailed 95\% credible interval).}

\item{Rhat}{logical. Do you wish to calculate the
Rhat convergence diagnostic using \code{\link[coda]{gelman.diag}}?
Fair warning: this can take a bit of time to run on many nodes/samples}

\item{neff}{logical. Do you wish to calculate the number of effective MCMC samples
using \code{\link[coda]{effectiveSize}}?
Fair warning: this can take a bit of time to run on many nodes/samples}

\item{mcse}{logical. Do you wish to calculate the
Monte Carlo standard error for the posterior mean and reported quantiles
using the \code{\link[mcmcse]{mcse}} and \code{\link[mcmcse]{mcse.q}} functions
(batch means method with batch size automatically calculated)?
Fair warning: this can take a bit of time to run on many nodes/samples}

\item{by_chain}{logical. Do you wish to calculate posterior summaries for each chain,
rather than for the aggregate across chains? Defaults to \code{FALSE}.
The arguments \code{Rhat}, \code{neff}, and \code{mcse} are ignored if \code{by_chain = TRUE}
and a warning will be returned}
}
\description{
Obtaining a summary of specific nodes is cumbersome from mcmc.list objects:
\code{\link[coda]{summary.mcmc.list}} is difficult to use for this purpose.
The central tendency and uncertainty measures are in different lists.
This function allows extracting a summary of the posteriors
associated with requested nodes, including several diagnostic quantities and measures of uncertainty.
}
\examples{
# load example mcmc.list
data(cjs, package = "postpack")

# calculate posterior summaries for the "p" nodes
# ("p[1]" doesn't exist in model)
post_summ(cjs, "p")

# do this by chain
post_summ(cjs, "p", by_chain = T)

# calculate Rhat and Neff diagnostic summaries as well
# multiple node names too
post_summ(cjs, c("b0", "p"), Rhat = T, neff = T)

# calculate Monte Carlo SE for mean and quantiles, with rounding
post_summ(cjs, "p", mcse = T, digits = 3)

# summarize different quantiles: median and central 80\%
post_summ(cjs, "p", probs = c(0.5, 0.1, 0.9))

}
\seealso{
\code{\link{match_params}}, \code{\link[coda]{gelman.diag}},
  \code{\link[coda]{effectiveSize}}, \code{\link[mcmcse]{mcse}}, \code{\link[mcmcse]{mcse.q}}
}
