% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_regex_lock}
\alias{rm_regex_lock}
\title{Remove the symbols that lock in a string for matching}
\usage{
rm_regex_lock(params)
}
\arguments{
\item{params}{a character vector on to which a carrot and dollar sign
will be removed from if present}
}
\description{
To ensure that a regular expression will match exactly,
it's necessary to specify so - this function undoes the work of
\code{\link{ins_regex_lock}} - currently just for prettier printing of node names
}
\note{
this is NOT a function users will generally use.
}
