% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_post.R
\name{summ_post}
\alias{summ_post}
\title{Obtain a posterior summary}
\usage{
summ_post(post, p = NULL, p_summ = c(0.5, 0.025, 0.975),
  prettify = F, rnd = NULL, bgr = F, ess = F)
}
\arguments{
\item{post}{an object of class \code{mcmc.list} from which to filter out the node
identified by \code{p}}

\item{p}{a character vector of length 1. Passed to \code{stringr::str_detect()},
so can (and perhaps should) be a regular expression.}

\item{p_summ}{numeric vector passed to \code{StatonMisc::summ()}.
The posterior percentiles you want to see. Defaults to \code{p_summ = c(0.5, 0.025, 0.975)}.}

\item{prettify}{logical passed to \code{StatonMisc::summ()}.
Do you wish to add commas to demarcate thousands points?
Coerces numerics to characters.}

\item{rnd}{numeric vector passed to \code{StatonMisc::summ()}.
The digits of \code{round()}.}

\item{bgr}{logical. Do you wish to calculate the
effective sample size using \code{coda:effectiveSize(post)}?
Fair warning: this can take a bit of time to run on many nodes/samples}

\item{ess}{logical. Do you wish to calculate the
effective sample size using \code{coda:effectiveSize(post)}?
Fair warning: this can take a bit of time to run on many nodes/samples}
}
\description{
Oftentimes you want to obtain the summary of just one node at a time.
\code{summary(post)} if \code{post} is of class \code{mcmc.list} is clunky.
The central tendency and uncertainty measures are in different lists.
This function allows extractinging a summary of the posteriors associated with requested nodes.
}
