% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_post.R
\name{thin_post}
\alias{thin_post}
\title{Thin a mcmc.list object}
\usage{
thin_post(post, thin_percent = 0.8)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{thin_percent}{numeric vector of length 1 and between the interval
(0,1): what fraction of the samples do you wish to remove from each
chain?}
}
\value{
an object of class \code{mcmc.list}
}
\description{
Systematically remove iterations from each chain of an
objective of class \code{mcmc.list}. Thinning after running the model is most often done during
  developement of post-processing code, so the calculations take
  less time.
}
\details{
the samples will be removed at as evenly spaced intervals
  as possible, however, this is not perfect. It is therefore recommended
  to use the full posterior for output, put this should be fine for
  most development. It can be improved by finding a better thinning rule:
  one that gives a consistent interval between all retained samples.
}
