% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Create a traceplot for a single desired node}
\usage{
trace_plot(post, param, keep_percent = 0)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{param}{a character vector of length 1.
Should be a node reference to a single element in the model. E.g., \code{"pi[1]"}, not \code{"pi"}.}

\item{keep_percent}{a numeric vector of length one and scaled between 0 and 1.
Percent of samples you'd like to keep for traceplotting and passed to \code{\link{post_thin}}.}
}
\description{
Create a traceplot for a single desired node
}
\note{
If saving as a pdf, these files can get very large with many samples and render slowly.
  The \code{keep_percent} argument is intended to help with this by thinning the chains at quasi-evenly spaced intervals.
}
