% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{write_model}
\alias{write_model}
\title{Export BUGS model from function to file}
\usage{
write_model(fun, file)
}
\arguments{
\item{fun}{a function object containing BUGS model code}

\item{file}{a character vector of length 1: the name of the file to write to}
}
\description{
Performs the same basic function as \code{\link[R2OpenBUGS]{write.model}}
}
\details{
This function performs the same basic function as \code{\link[R2OpenBUGS]{write.model}},
    but with slightly better output (scientific notation, spacing, etc.). The main reason it was created
    for use in this package was to remove the need for the \code{R2OpenBUGS} package, as this was
    the only function the author ever used from it.
}
