% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_format.R
\name{array_format}
\alias{array_format}
\title{Convert a vector to the array structure used in model}
\usage{
array_format(v)
}
\arguments{
\item{v}{a vector with names indicating its elements' appropriate placement in a new array.
See the details for more about what this means.}
}
\description{
This function unvectorizes summarized output and creates an array
of the appropriate dimensions based on the element names.
}
\details{
For example, you have a AxB matrix \code{m} in your JAGS model, and you would like to
  create an object in R that stores the posterior means in the same location as
  in the JAGS model. If A and B are small, it is simple to create this matrix "by-hand".
  However, if there are also dimensions C, D, and E, it becomes much more difficult.
  Up to 10 dimensions are currently supported.
}
\examples{
\dontrun{
array_format(post_summ(post, "^m[")["mean",])
}

}
