% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{Create a density for a single desired node}
\usage{
density_plot(post, p_one)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{p_one}{a character vector of length == 1. Should be a node reference
to a single element in the model. E.g., \code{"pi[1]"}, not \code{"pi"}.}
}
\description{
Create a density for a single desired node
}
