% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{Create a density for a single desired node}
\usage{
density_plot(post, param, show_diags = "if_poor_Rhat")
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{param}{a character vector of length == 1. Should be a node reference
to a single element in the model. E.g., \code{"pi[1]"}, not \code{"pi"}.}

\item{show_diags}{a character vector of length == 1. Must be one of
\code{"always"}, \code{"never"}, \code{"if_poor_Rhat"}. Defaults to
\code{"if_poor_Rhat"}, which will display the Rhat and effective MCMC samples
if the Rhat statistic is greater than 1.1}
}
\description{
Create a density for a single desired node
}
