% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{diag_plots}
\alias{diag_plots}
\title{Create MCMC diagnostic plots for nodes of interest}
\usage{
diag_plots(
  post,
  p,
  ext_device = FALSE,
  show_diags = "if_poor_Rhat",
  layout = "auto",
  dims = "auto",
  keep_percent = 1,
  save = FALSE,
  file = NULL,
  auto_escape = TRUE
)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{p}{a character vector of with length >= 1 specifying the nodes to plot. Passed to \code{\link{match_p}},
so can (and sometimes should) be a regular expression.}

\item{ext_device}{logical. Do you wish to have an external device open to display the diagnostics?
\code{TRUE} will create a new plotting device using the OS-specific function.
Defaults to \code{FALSE}. Requires the \code{\link[StatonMisc]{ext_device}} function.}

\item{show_diags}{a character vector of length == 1. Must be one of
\code{"always"}, \code{"never"}, \code{"if_poor_Rhat"}. Defaults to
\code{"if_poor_Rhat"}, which will display the Rhat and effective MCMC samples
if the Rhat statistic is greater than 1.1}

\item{layout}{a character vector specifying \code{"ROWSxCOLUMNS"} of parameter diagnostics.
For example, \code{"4x1"} has 4 rows and 1 column of parameter diagnostics.
Defaults to \code{"auto"}, which selects between \code{"1x1"}, \code{"2x1"}, \code{"4x1"}, \code{"4x2"}, and \code{"5x3"}.}

\item{dims}{a character vector specifying the \code{"HEIGHTxWIDTH"} dimensions
of the plotting device, in inches.
For example, \code{"5x7"} would create a 5 inch tall and 7 inch wide plotting device.
Defaults to \code{"auto"}, which selects the dimensions that look nice when \code{layout = "auto"}
as well.}

\item{keep_percent}{a numeric vector of length one and scaled between 0 and 1.
Percent of samples you'd like to keep for traceplotting and passed to \code{\link{post_thin}}.}

\item{save}{logical. Do you wish to save the diagnostic plots in a PDF file? If so,
specify \code{file = "example.pdf"} as well. Defaults to \code{FALSE}.}

\item{file}{character vector of length 1. The file name, which
must include the \code{".pdf"} extension. Saved to working directory by default,
but can recieve an absolute or relative file path here as well}

\item{auto_escape}{logical. \code{FALSE} will treat \code{"["} and \code{"]"}
as regular expression syntax (unless explicitly escaped by user),
\code{TRUE} will treat these symbols as plain text to be matched.
It is generally recommended to keep this as \code{TRUE} (the default),
unless you are performing complex regex searches that require the
\code{"["} and \code{"]"} symbols to be special characters}
}
\description{
Allows quick visualization of posterior density and traceplots,
BOTH separated by chain, for the desired nodes of interest. Includes the
ability to plot in the RStudio graphics device, an external device (OS-indepenent),
or a PDF file. Further, with the auto settings, the dimensions of the
plotting device scales to the job needed.
}
\note{
If saving as a pdf, these files can get very large with many samples and render slowly.
  The \code{keep_percent} argument is intended to help with this by thinning the chains at quasi-evenly spaced intervals.
}
\seealso{
\code{\link{match_p}}
}
