% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{diag_plots}
\alias{diag_plots}
\title{Create MCMC diagnostic plots for nodes of interest}
\usage{
diag_plots(post, p, layout = "auto", dims = "auto", ext_device = F,
  thin_percent = 0, save = F, file = NULL, warn = F)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{p}{a character vector of with length >= 1 specifying the nodes to plot. Passed to \code{\link{match_p}},
so can (and sometimes should) be a regular expression.}

\item{layout}{a character vector specifying \code{"ROWSxCOLUMNS"} of parameter diagnostics.
For example, \code{"4x1"} has 4 rows and 1 column of parameter diagnostics.
Defaults to \code{"auto"}, which selects between \code{"1x1"}, \code{"4x1"}, and \code{"5x3"}.}

\item{dims}{a character vector specifying the \code{"HEIGHTxWIDTH"} dimensions
of the plotting device, in inches.
For example, \code{"5x7"} would create a 5 inch tall and 7 inch wide plotting device.
Defaults to \code{"auto"}, which selects the dimensions that look nice when \code{layout = "auto"}
as well.}

\item{ext_device}{logical. Do you wish to have an external device open to display the diagnostics?
\code{TRUE} will create a new plotting device using the OS-specific function.
Defaults to \code{FALSE}. Requires the \code{\link[StatonMisc]{ext_device}} function.}

\item{thin_percent}{if you wish to remove some fraction of the samples before traceplotting, do that here.
The thinning will occur systematically, i.e., at evenly-spaced intervals.
This is sometimes helpful when saving a pdf with many samples/nodes, which can render slowly.
Defaults to \code{0}, i.e., no thinning.}

\item{save}{logical. Do you wish to save the diagnostic plots in a PDF file? If so,
specify \code{file = "example.pdf"} as well. Defaults to \code{FALSE}.}

\item{file}{character vector of length 1. The file name, which
must include the \code{".pdf"} extension. Saved to working directory by default,
but can recieve an absolute or relative file path here as well}

\item{warn}{logical. Do you wish to receive warnings about
possibly unexpected behavior regarding regex matching? Defaults to \code{FALSE}}
}
\description{
Allows quick visualization of posterior density and traceplots,
BOTH separated by chain, for the desired nodes of interest. Includes the
ability to plot in the RStudio graphics device, an external device (OS-indepenent),
or a PDF file. Further, with the auto settings, the dimensions of the
plotting device scales to the job needed.
}
\seealso{
\code{\link{match_p}}
}
