% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_index}
\alias{drop_index}
\title{Extract the base of a node name}
\usage{
drop_index(params)
}
\arguments{
\item{params}{character vector containing node names}
}
\description{
Removes square brackets, numbers, and commas that represent
the index of the node element in question. Returns just the
node name.
}
\note{
this is NOT a function users will generally use.
}
