% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Get the names of saved nodes}
\usage{
get_params(post, type = "base")
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{type}{character vector of length == 1. Currently accepted options are
\code{"unique"} and \code{"all"}. See below.}
}
\value{
if \code{type = "unique"}, then the unique node names (i.e., without indices)
  will be returned. If \code{type = "all"}, then all node names, including the indices,
  will be returned.
}
\description{
Sometimes it's helpful to know what nodes are
stored in an object of class \code{mcmc.list}.
That's what this function does.
}
\seealso{
\code{\link{drop_index}}
}
