% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Get the names of saved nodes}
\usage{
get_params(post, type = "base_only")
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{type}{a character vector with length == 1; only two options are accepted.
Set to \code{type = "base_only"} (the default) if you wish to return only the unique node names (without indices).
Set to \code{type = "base_index"}  if you wish to return the node names with indices included.}
}
\value{
character vector with all node names stored in the object \code{post}, formatted as requested by \code{type}
}
\description{
Sometimes it's helpful to know what nodes are
stored in an object of class \code{mcmc.list}.
That's what this function does.
}
\examples{
 # load example mcmc.list
data(cjs, package = "postpack")

# get only node names, no indices (default)
get_params(cjs, type = "base_only")

# get indices too, where applicable
get_params(cjs, type = "base_index")
}
