% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mat.R
\name{id_mat}
\alias{id_mat}
\title{Quickly extract the chain and iteration IDs}
\usage{
id_mat(post)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}
}
\value{
a matrix with columns \code{"CHAIN"} and \code{"ITER"}
}
\description{
It is sometimes useful to be able to add these characteristics to a 
matrix of derived quantities, which then allows coercion to a \code{mcmc.list} 
object, which can then be binded (e.g., \code{\link{post_bind}}) to the 
more complete samples.
}
