% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ins_regex_bracket}
\alias{ins_regex_bracket}
\title{Insert escapes on regex brackets}
\usage{
ins_regex_bracket(params)
}
\arguments{
\item{params}{Node names.}
}
\value{
A character vector with all brackets escaped. For example,
\code{"a[1]"} becomes \verb{"a\\\[1\\\]"}
}
\description{
Insert escapes on regex brackets
}
\details{
Searches the contents of a string for the occurrence of a
square bracket or two, and inserts the necessary escapes for pattern matching
via regular expressions.
}
\note{
This is \strong{not} a function users will generally use directly.
}
