% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ins_regex_bracket}
\alias{ins_regex_bracket}
\title{Insert a brackets as regex version}
\usage{
ins_regex_bracket(string)
}
\arguments{
\item{string}{a character vector}
}
\description{
Insert a brackets as regex version
}
