% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ins_regex_lock}
\alias{ins_regex_lock}
\title{Insert the symbols to lock in a string for matching}
\usage{
ins_regex_lock(params)
}
\arguments{
\item{params}{a character vector on to which a carrot and dollar sign
will be pasted to if not already present}
}
\description{
To ensure that a regular expression will match exactly,
it's necessary to specify so - which is what this function does.
}
\note{
this is NOT a function users will generally use.
}
