% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_out.R
\name{list_out}
\alias{list_out}
\title{List vector elements in a nice format}
\usage{
list_out(x, final = NULL, per_line = 1e+06, wrap = NULL, indent = NULL)
}
\arguments{
\item{x}{a vector object, will be coerced to a character class}

\item{final}{character vector of length == 1: word that will separate the final element in the list from others.
See the examples}

\item{per_line}{numeric vector of length == 1: number of elements printed per line. See the examples}

\item{wrap}{character vector of length == 1: optional character to wrap around each element,
e.g., quotation marks}

\item{indent}{character vector of length == 1: anything you wish to place in front of the first
element on each line. See the examples}
}
\value{
a character vector of length == 1, ready to be passed to
  \code{\link[base]{stop}}, \code{\link[base]{warning}}, or \code{\link[base]{cat}}, for example.
}
\description{
Oftentimes in my functions I like to list out the accepted arguments if
the user enters one that is not accepted. This function faciliates that.
}
\examples{
list_out(x = letters[1:10])
cat(list_out(letters[1:26], wrap = "'", per_line = 4, final = "or", indent = "  "))
}
