% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2mcmclist.R
\name{matrix2mcmclist}
\alias{matrix2mcmclist}
\title{Coerce a matrix to an mcmc.list}
\usage{
matrix2mcmclist(post_mat)
}
\arguments{
\item{post_mat}{a posterior model object, with samples in matrix format.
The matrix can contain any number of nodes, but it must contain
\code{"CHAIN"} and \code{"ITER"} are columns.}
}
\description{
Does as the name implies. Requires that the \code{"CHAIN"} and \code{"ITER"} are columns.
These can be obtained from the \code{filter_post()} function.
}
