% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mytitle}
\alias{mytitle}
\title{Add a title between two figures}
\usage{
mytitle(text)
}
\arguments{
\item{text}{A character vector length == 1. The text string to
include as a title.}
}
\description{
Used by \code{\link[=diag_plots]{diag_plots()}} to place a common
title over top of two figures: one density and one trace
for a given node.
}
\note{
This is \strong{not} a function users will generally use directly.
}
