% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_bind.R
\name{post_bind}
\alias{post_bind}
\title{Combine two mcmc.list objects}
\usage{
post_bind(post1, post2, dup_id = "_p2")
}
\arguments{
\item{post1}{an object of class \code{mcmc.list}}

\item{post2}{an object of class \code{mcmc.list}}

\item{dup_id}{a character vector of length == 1. If any
node names are duplicated in post2, what should be appended to the
end of the node names in the output? Defaults to \code{"_p2"}}
}
\description{
Bind together two mcmc.list objects, possibly from the same or different models.
The two posterior objects must have equal chains and iterations per chain.
}
\details{
The two lists are coerced to matricies, \code{cbind()}-ed, and
  converted back to to an \code{mcmc.list} using \code{matrix2mcmclist}.
  The chains, thinning, and iterations, burn-in settings in the output object
  will remain as for \code{post1}.
}
