% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dim.R
\name{post_dim}
\alias{post_dim}
\title{Obtain MCMC dimensions from an mcmc.list}
\usage{
post_dim(post)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}
}
\value{
vector with named elements
\itemize{
  \item \code{burn} - The burn-in period (per chain)
  \item \code{post_burn} - The post-burn-in period (per chain)
  \item \code{thin} - The thinning interval post-burn-in
  \item \code{chains} - The number of chains
  \item \code{saved} - The number of saved samples across all chains
  \item \code{nodes} - The number of nodes with MCMC samples
}
}
\description{
Quickly query the number of burn-in samples, post-burnin, thinning,
number of chains, etc. from an object of class \code{mcmc.list}
}
\note{
if the \code{post} object was thinned after MCMC completed
  using \code{\link{post_thin}}, then these dimensions will be improperly calculated
  (currently).
}
