% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_subset.R
\name{post_subset}
\alias{post_subset}
\title{Extract samples from specific nodes}
\usage{
post_subset(
  post,
  params,
  matrix = FALSE,
  iters = FALSE,
  chains = FALSE,
  auto_escape = TRUE
)
}
\arguments{
\item{post}{A \code{\link[coda:mcmc.list]{mcmc.list}} object.}

\item{params}{A character vector of with length >= 1 specifying the nodes to extract from \code{post}.
Passed to \code{\link[=match_params]{match_params()}} so is matched using regular expressions.}

\item{matrix}{Logical. If \code{TRUE}, the output will be returned as a \code{\link[base:matrix]{matrix}}.
Defaults to \code{FALSE}, in which case the \code{\link[coda:mcmc.list]{mcmc.list}} class will be retained.}

\item{iters}{Logical. Do you wish to retain the iteration number of each sample if \code{matrix = TRUE}? Not used otherwise.
Defaults to \code{FALSE}.}

\item{chains}{Logical. Do you wish to retain the chain number of each sample if \code{matrix = TRUE}? Not used otherwise.
Defaults to \code{FALSE}.}

\item{auto_escape}{Logical. \code{FALSE} will treat \code{"["} and \code{"]"}
as special regular expression characters (unless explicitly escaped by user),
\code{TRUE} will treat these symbols as plain text to be matched.
It is generally recommended to keep this as \code{TRUE} (the default),
unless you are performing complex regex searches that require the
\code{"["} and \code{"]"} symbols to be special characters.}
}
\description{
Subsets a smaller portion from a \code{\link[coda:mcmc.list]{mcmc.list}} object
corresponding only to the node(s) requested.
}
\examples{
# load example mcmc.list
data(cjs, package = "postpack")

# create mcmc.list with all nodes that contain "B0"
x1 = post_subset(cjs, "B0")

# create mcmc.list with all nodes that contain "b" or "B"
x2 = post_subset(cjs, c("b", "B"))

# perform the subset and return a matrix as output, while retaining the chain ID
x3 = post_subset(cjs, "B0", matrix = T, chain = T)
}
\seealso{
\code{\link{match_params}}
}
