% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_subset.R
\name{post_subset}
\alias{post_subset}
\title{Extract a subset of nodes from a \code{mcmc.list}}
\usage{
post_subset(post, p, matrix = FALSE, iters = F, chains = F,
  warn = F)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{p}{a character vector of with length >= 1 specifying the nodes to extract. Passed to \code{\link{match_p}},
so can (and sometimes should) be a regular expression.}

\item{matrix}{logical. If \code{TRUE}, the subsetted output will be returned as a matrix.
Defaults to \code{FALSE}, in which case the class \code{mcmc.list} will be retained.}

\item{iters}{logical. Do you wish to retain the iteration number if \code{matrix = TRUE}? Not used otherwise.
Defaults to \code{FALSE}.}

\item{chains}{logical. Do you wish to retain the chain number if \code{matrix = TRUE}? Not used otherwise.
Defaults to \code{FALSE}.}

\item{warn}{logical. Do you wish to receive warnings about
possibly unexpected behavior regarding regex matching?}
}
\description{
Subsets a smaller portion from the joint posterior
corresponding only to the node(s) requested.
}
\seealso{
\code{\link{match_p}}
}
