% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_summ.R
\name{post_summ}
\alias{post_summ}
\title{Obtain a posterior summary of specific nodes}
\usage{
post_summ(
  post,
  p,
  rnd = NULL,
  p_summ = c(0.5, 0.025, 0.975),
  Rhat = FALSE,
  ess = FALSE,
  mcse = FALSE,
  auto_escape = TRUE
)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{p}{a character vector of with length >= 1 specifying the nodes to summarize.
Passed to \code{\link{match_p}}, so can (and sometimes should) be a regular expression.}

\item{rnd}{numeric vector passed to \code{\link[StatonMisc]{summ}}.
The digits of \code{round}.}

\item{p_summ}{numeric vector passed to \code{\link[StatonMisc]{summ}}.
The posterior percentiles you want to have summarized, Defaults to \code{p_summ = c(0.5, 0.025, 0.975)}.}

\item{Rhat}{logical. Do you wish to calculate the
Rhat convergence diagnostic using \code{\link[coda]{gelman.diag}}?
Fair warning: this can take a bit of time to run on many nodes/samples}

\item{ess}{logical. Do you wish to calculate the
effective sample size using \code{\link[coda]{effectiveSize}}?
Fair warning: this can take a bit of time to run on many nodes/samples}

\item{mcse}{logical. Do you wish to calculate the
Monte Carlo standard error for the posterior mean and reported quantiles
using the \code{\link[mcmcse]{mcse}} and \code{\link[mcmcse]{mcse.q}} functions
(batch means method with batch size automatically calculated)?
Fair warning: this can take a bit of time to run on many nodes/samples}
}
\description{
Obtaining a summary of specific nodes is cumbersome from mcmc.list objects:
\code{summary(post)} if \code{post} is of class \code{mcmc.list} is clunky.
The central tendency and uncertainty measures are in different lists.
This function allows extractinging a summary of the posteriors
associated with requested nodes.
}
\seealso{
\code{\link{match_p}}, \code{\link[StatonMisc]{summ}}, \code{\link[coda]{gelman.diag}},
  \code{\link[coda]{effectiveSize}}, \code{\link[mcmcse]{mcse}}, \code{\link[mcmcse]{mcse.q}}
}
