% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_summ.R
\name{post_summ}
\alias{post_summ}
\title{Obtain a posterior summary of specific nodes}
\usage{
post_summ(
  post,
  params,
  rnd = NULL,
  p_summ = c(0.5, 0.025, 0.975),
  Rhat = FALSE,
  ess = FALSE,
  mcse = FALSE,
  by_chain = F,
  auto_escape = TRUE
)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{params}{a character vector of with length >= 1 specifying the nodes to summarize.
Passed to \code{\link{match_params}}, so can (and sometimes should) be a regular expression.}

\item{rnd}{numeric vector controlling rounding of summaries.
Passed to the \code{digits} argument of \code{\link[base]{round}}.
Defaults to \code{NULL}, which produces no rounding.}

\item{p_summ}{numeric vector with the posterior percentiles you wish to have summarized.
Defaults to \code{p_summ = c(0.5, 0.025, 0.975)}.}

\item{Rhat}{logical. Do you wish to calculate the
Rhat convergence diagnostic using \code{\link[coda]{gelman.diag}}?
Fair warning: this can take a bit of time to run on many nodes/samples}

\item{ess}{logical. Do you wish to calculate the
effective sample size using \code{\link[coda]{effectiveSize}}?
Fair warning: this can take a bit of time to run on many nodes/samples}

\item{mcse}{logical. Do you wish to calculate the
Monte Carlo standard error for the posterior mean and reported quantiles
using the \code{\link[mcmcse]{mcse}} and \code{\link[mcmcse]{mcse.q}} functions
(batch means method with batch size automatically calculated)?
Fair warning: this can take a bit of time to run on many nodes/samples}

\item{by_chain}{logical. Do you wish to calculate posterior summaries for each chain,
rather than for the aggregate across chains? Defaults to \code{FALSE}.
The arguments \code{Rhat}, \code{ess}, and \code{mcse} are ignored if \code{by_chain = TRUE}
and a warning will be returned}
}
\description{
Obtaining a summary of specific nodes is cumbersome from mcmc.list objects:
\code{summary(post)} if \code{post} is of class \code{mcmc.list} is clunky.
The central tendency and uncertainty measures are in different lists.
This function allows extractinging a summary of the posteriors
associated with requested nodes.
}
\seealso{
\code{\link{match_params}}, \code{\link[coda]{gelman.diag}},
  \code{\link[coda]{effectiveSize}}, \code{\link[mcmcse]{mcse}}, \code{\link[mcmcse]{mcse.q}}
}
