% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_summ.R
\name{post_summ}
\alias{post_summ}
\title{Obtain posterior summaries and diagnostics of specific nodes}
\usage{
post_summ(
  post,
  params,
  digits = NULL,
  probs = c(0.5, 0.025, 0.975),
  Rhat = FALSE,
  neff = FALSE,
  mcse = FALSE,
  by_chain = F,
  auto_escape = TRUE
)
}
\arguments{
\item{post}{A \code{\link[coda:mcmc.list]{mcmc.list}} object.}

\item{params}{A character vector of with length >= 1 specifying the nodes to summarize from \code{post}.
Passed to \code{\link[=match_params]{match_params()}} so is matched using regular expressions.}

\item{digits}{A numeric vector with length == 1 controlling rounding of summaries.
Passed to \code{\link[base:Round]{base::round()}} and defaults to \code{NULL}, which produces no rounding.}

\item{probs}{A numeric vector specifying the posterior quantiles you wish to have summarized.
Passed to \code{\link[stats:quantile]{stats::quantile()}}.
Defaults to \code{probs = c(0.5, 0.025, 0.975)} (i.e., median and equal-tailed 95 percent credible interval).}

\item{Rhat}{Logical. Do you wish to calculate the
Rhat convergence diagnostic using \code{\link[coda:gelman.diag]{coda::gelman.diag()}}?
Fair warning: this can take a bit of time to run on many nodes/samples.}

\item{neff}{Logical. Do you wish to calculate the number of effective MCMC samples
using \code{\link[coda:effectiveSize]{coda::effectiveSize()}}?
Fair warning: this can take a bit of time to run on many nodes/samples.}

\item{mcse}{Logical. Do you wish to calculate the
Monte Carlo standard error for the posterior mean and reported quantiles
using the \code{\link[mcmcse:mcse]{mcmcse::mcse()}} and \code{\link[mcmcse:mcse.q]{mcmcse::mcse.q()}} functions
(batch means method with batch size automatically calculated)?
Fair warning: this can take a bit of time to run on many nodes/samples.}

\item{by_chain}{Logical. Do you wish to calculate posterior summaries for each chain
rather than for the aggregate across chains? Defaults to \code{FALSE}.
The arguments \code{Rhat}, \code{neff}, and \code{mcse} are ignored if \code{by_chain = TRUE}
and a warning will be returned.}
}
\value{
A \code{\link[base:matrix]{matrix}} object with summary statistics as rows and nodes as columns.
If \code{by_chain = TRUE}, an \code{\link[base:array]{array}} is returned instead.
}
\description{
Allows rapid calculation of summaries and diagnostics from \strong{specific nodes}
stored in \code{\link[coda:mcmc.list]{mcmc.list}} objects.
}
\examples{
# load example mcmc.list
data(cjs, package = "postpack")

# calculate posterior summaries for the "p" nodes
# ("p[1]" doesn't exist in model)
post_summ(cjs, "p")

# do this by chain
post_summ(cjs, "p", by_chain = TRUE)

# calculate Rhat and Neff diagnostic summaries as well
# multiple node names too
post_summ(cjs, c("b0", "p"), Rhat = TRUE, neff = TRUE)

# calculate Monte Carlo SE for mean and quantiles, with rounding
post_summ(cjs, "p", mcse = TRUE, digits = 3)

# summarize different quantiles: median and central 80\%
post_summ(cjs, "p", probs = c(0.5, 0.1, 0.9))
}
\seealso{
\code{\link[=match_params]{match_params()}}, \code{\link[coda:gelman.diag]{coda::gelman.diag()}}, \code{\link[coda:effectiveSize]{coda::effectiveSize()}}, \code{\link[mcmcse:mcse]{mcmcse::mcse()}}, \code{\link[mcmcse:mcse.q]{mcmcse::mcse.q()}}
}
