% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_regex_bracket}
\alias{rm_regex_bracket}
\title{Remove escapes on regex brackets}
\usage{
rm_regex_bracket(string)
}
\arguments{
\item{string}{a character vector}
}
\description{
Remove escapes on regex brackets
}
\details{
searches the contents of a string for the occurrence of a
 square bracket or two (that has been escaped), and removes the escaping
 that was necessary for regular expression matching
}
\note{
this is NOT a function users will generally use.
}
