% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Create a traceplot for a single desired node}
\usage{
trace_plot(post, p = NULL, thin_percent = 0)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{thin_percent}{a numeric vector of length one and scaled between 0 and 1.
of samples you'd like to remove before traceplotting. If saving as a pdf,
these files can get very large with many samples and render slowly.
Samples are thinned systematically (i.e., at evenly-spaced intervals).}

\item{p_one}{a character vector of length 1.
Should be a node reference to a single element in the model. E.g., \code{"pi[1]"}, not \code{"pi"}.}
}
\description{
Create a traceplot for a single desired node
}
